update errore controlloStorico

+++ Source code/Aeroporto/src/main/java/com/pack/aeroporto/entity/Posto.java
@@ -26,11 +26,11 @@ public class Posto {
 		this.prenotato = prenotato;
 	}
 	
-	public int getNumero() {
+	public Long getNumero() {
 		return numero;
 	}
 	
-	public void setNumero(int numero) {
+	public void setNumero(Long numero) {
 		this.numero = numero;
 	}
 }
